refactor(commands): extract diagnostic commands into diagnostics.rs
